🛠️ refactor: 코드 리펙토링
- 레디스 사용시 일치 하지 않은 에러처리를 통일

🔗 Resolves: #42

+++ backend/JiShop/src/main/java/com/jishop/category/dto/SubCategory.java
@@ -3,6 +3,6 @@ package com.jishop.category.dto;
 public record SubCategory(
         Long id,
         String name,
-        int productCount
+        long productCount
 ) {
 }
